drm_atomic: Allow to create atomic context w/o drmprime video plane
This is to improve the experience when running with default settings
on a driver that doesn't have any overlay planes, but still supports
drm atomic. Since the drmprime video plane is set to pick an overlay
plane by default it would fail on these drivers due to not being able
to create any atomic context. Users with such cards had to specify
--drm-video-plane-id manually to some bogus value (it's not used after
all), or disable atomic with --drm-atomic=no to get it to work.

The "video" plane is only ever used by the drmprime-drm hwdec interop,
which is not used at all in the typical usecase where everything is
actually rendered on to the "OSD" plane using EGL, so having an atomic
context without the "video" plane should be fine most of the time.