demux: fix some theoretical UB with no impact
If the number of chapters is 0, the chapter list can be NULL. clang
complains that we pass NULL to qsort(). This is yet another pointless UB
that exists for no reason. Why the fuck? Why can't I just pass NULL?
What retard in clang decided that this is bad? Fucking shit fuck you
fucking retards THIS IS FUCKING COMPLETELY UNNECESSARY AND A WASTE OF
FUCKING TIME GOD DAMN FUCK YOU IDIOTS