drm: rename plane options to better, invariant, names
This commit bumps the libmpv version to 1.102

drm-osd-plane -> drm-render-plane
drm-video-plane -> drm-drmprime-video-plane
drm-osd-size -> drm-render-size

The new names are invariant whether or not hwdec_drmprime_drm is being
used or not. The original naming was very confusing, as when doing
regular rendering (swdec or vaapi) the video would be displayed on the
"OSD plane", and the "Video plane" would remain unused.

"render plane", as in the plane where OpenGL renders it's thing
whether it be video + OSD or just OSD, and "drmprime video plane", as
in the plane used for hwdec video imported via drmprime, is much more
correct.