context_drm_egl: Fix display of OSD and other when paused
The implementation of triple buffering has a problem when it is no
longer being called for every vsync: It needs to be called twice to
even queue the latest frame for presentation.

Work around this by switching to what is effectively double-buffered
when the player is paused.

Problem/TODO: Doesn't work properly for coverart, which is also
render-seldom, even when the player is not paused.