context_drm_egl: Drain the swapchain when paused/displaying cover art
The way triple buffering is implemented in context_drm_egl caused some
problems when paused or for still images like cover art. This is
because the page flip for the currently drawn image was only being
queued on the subsequent call to swap_buffers.

To fix this schedule an extra page flip when paused/displaying still
images, so that output doesn't lag behind user input. This causes
swap_buffers to wait for two vsyncs in some cases, but this doesn't
seem to cause any major issues.