context_drm_egl: Always use deferred wait
This effectively limits the depth to one less than what is the maximum amount of
buffers you can have locked at the same time with GBM. On Mesa, which at the
time of writing allows to lock up to 4 buffers, this means triple buffering or
--swapchain-depth=2. I suspect that the reason Mesa has chosen a max size of 4
buffers here is precisely because it should be possible to do triple buffering
in this fashion, while always keeping one buffer unlocked, so that you can start
drawing the next frame immediately and defer the wait for page flip.

Although, since gbm_surface_has_free_buffers() can only tell us a
boolean value and not how many buffers we have left, we are forced to
do this contortionist dance where we first overshoot until
gbm_surface_has_free_buffers() reports 0, followed by immediately
waiting so we can free a buffer, to be able to get the deferred wait
rolling.