sub: implement sub-scale for DVD/Vobsubs
Use libswscale to scale DVD/Vobsubs, as what has been done in the
DVD/Vobsubs blurring code.

The subjective part is the position of the subtitle after scaling. I
choose to fix the bottom-center point of the subtitle bitmap, because
commonly subtitles are placed near the bottom of the video. This is also
consistent with the behavior when a text subtitle is scaled.

Fixes #6417