Add presentation feedback for wayland [WIP]
This requires having update_vsync_timing_after_swap be called
externally when external renderloop is enabled. So we rename it
update_vsync_timing_after_swap_external and make it externally
callable.

Currently we just use mp_time_us in the presentation callback, as that
should already be better. Later we might want to also use the various
variables passed to pts_presented.