demux_lavf: also fix cachec seeking with large codec delay
Fixes the same thing as the previous commit did with demux_mkv. I'm not
sure if this is correct or a good idea (well, it works with my sample
file).

In fact, this shit looks generally a bit gross, why the fuck is our
wrapper compensating for the seek delay manually? (But I checked that
this is really needed. I guess libavformat is really into making people
do gross, nasty shit.)