demux: fix index resizing with wraparound
Of course this code was extremely buggy. It didn't consider the case
when the list needs to be resized, and the queue wraps around the end of
the array. It actually gets triggered relatively rarely, but when it
does, it will trigger UB big time. Another mistake is that index0 wasn't
wrapped, and the QUEUE_INDEX_ENTRY macro didn't quote the idx parameter.

I didn't really test this, and there will be no tests, since this is
just an ad-hoc reinvented wheel welded to its location of use.

The loop could be a single memcpy(), but let's not overdo this.