demux: fix stream recording when --cache-on-disk is used
Passed a packet that had is_cache set to code which didn't expect it.
Blatant UB.

Fix this by passing the packet to the muxer before it's put in the
cache. Move the glue code for muxing into a separate function to reduce
the noise.

Some minor details change due to this, but probably nothing important.