vo_drm [WIP]: make VT switching work w/o drmDropMaster
Since drmSetMaster and drmDropMaster are broken to the point of being
useless, we instead use the ancient method of just closing or FD to
/dev/dri/cardX.

Seems to work.

TODO: cleanup. and the same for gpu-context=drm (hopefully we can just
keep /dev/dri/renderDXXX open throughout)

vo_drm: misc fixes

shoudl probably apply most of this already in the n-buffering place